complete bert convert
